+++ src/main.cpp
@@ -138,6 +138,8 @@ int dispatch(Settings settings, const std::vector<std::string>& args) {
       std::cout << "error updating executable" << std::endl;
       return 1;
     }
+    std::cout << "Update installed. Restarting setup... " << std::endl
+              << std::endl;
     // then start the setup of the new one (create a new process)
     cmd = "\"" + args.at(2) + "\" setup";
     create_win_process(cmd);